Create the dktest container

+++ ssh_docker/add_ssh_key.sh
@@ -1,9 +1,13 @@
 #!/bin/sh
-rm -f /tmp/dimail_api_test_id_rsa /tmp/dimail_api_test_id_rsa.pub
+NAME=$1
+
+if [ -f /tmp/dimail_${NAME}_id_rsa -o -f /tmp/dimail_${NAME}_id_ras.pub ]; then
+	rm -f /tmp/dimail_api_test_id_rsa /tmp/dimail_api_test_id_rsa.pub
+fi
 
 ssh-keygen -t rsa -b 4096 -q -N "" \
-  -f /tmp/dimail_api_test_id_rsa \
+  -f /tmp/dimail_${NAME}_id_rsa \
   -C "tu@dimail"
 
-cat /tmp/dimail_api_test_id_rsa.pub > ../oxtest/authorized_keys2
+cat /tmp/dimail_${NAME}_id_rsa.pub > ../${NAME}/authorized_keys2
